sql: move tableUpserter and fastTableUpserter to their own files
Release note: None
